+++ xerkonix-design-system/lib/src/tactile/tactile_primary_button.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 import 'tactile_tokens.dart';
 import 'tactile_type.dart';